Fix issue #12 (Fails to parse Wallet Journal) (#2)
Add a temporary constructor for the Station for citadels. This will allow to see Citadel purchases in the Wallet Transaction. Name for citadel is not resolved yet.

Note: should be replaced with a proper citadel system once it's available.
Also, should create an interface for the station & citadel in the WalletTransaction system.